Inline Filterable and Sortable modules into Declarative Filters
This flattens the methods that are exposed to controllers into a single mix. If anyone opens this class, they can see the method  signatures and documentation in one place.
